#24: Fixed review findings.
